djangoProject/resume_builder_django/requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap4==2025.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.0.14
django-crispy-forms==2.4
django-formtools==2.5.1
django-heroku==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
fonttools==4.58.5
lxml==6.0.0
mssql-django==1.5
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
pyodbc==5.2.0
pyphen==0.17.2
python-docx==1.2.0
pytz==2025.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
